refactor(cli): extract module parser owner
